HW3 - COMMAND LINE. DEBUGGING. ERROR HANDLING / FILE SYSTEM AND STREAMS
- refactoring: move functionality to different modules to achieve single responsibility and high cohesion
